fix: bug auth on createBook function

+++ controllers/book.js
@@ -6,7 +6,8 @@ exports.createBook = (req, res, next) => {
   delete req.body._id;
   delete req.body._userId;
   const book = new Book({
-    ...bookThing, // add user here !!
+    ...bookThing,
+    userId: req.auth.userId,
     imageUrl: `${req.protocol}://${req.get('host')}/images/${
       req.file.filename
     }`,
@@ -58,18 +59,18 @@ exports.modifyBook = (req, res, next) => {
   delete bookObject._userId;
   Book.findOne({ _id: req.params.id })
     .then((book) => {
-      // if (bookObject.userId != req.auth.userId) {
-      //   res.status(401).json({ message: 'Not authorized '})
-      // } else {
-      Book.updateOne(
-        { _id: req.params.id },
-        { ...bookObject, _id: req.params.id }
-      )
-        .then(() => {
-          res.status(200).json({ message: 'Modified!' });
-        })
-        .catch((error) => res.status(401).json({ error }));
-      // }
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized ' });
+      } else {
+        Book.updateOne(
+          { _id: req.params.id },
+          { ...bookObject, _id: req.params.id }
+        )
+          .then(() => {
+            res.status(200).json({ message: 'Modified!' });
+          })
+          .catch((error) => res.status(401).json({ error }));
+      }
     })
     .catch((error) => {
       res.status(400).json({ error });
@@ -79,20 +80,20 @@ exports.modifyBook = (req, res, next) => {
 exports.deleteBook = (req, res, next) => {
   Book.findOne({ _id: req.params.id })
     .then((book) => {
-      // if (thing.userId !=  req.auth.userId) {
-      //   res.status(401).json({ message: 'Not authorized!'})
-      // } else {
-      const filename = book.imageUrl.split('/images/')[1];
-      fs.unlink(`images/${filename}`, () => {
-        Book.deleteOne({ _id: req.params.id })
-          .then(() => {
-            res.status(200).json({ message: 'Deleted!' });
-          })
-          .catch((error) => {
-            res.status(401).json({ error });
-          });
-      });
-      // }
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized!' });
+      } else {
+        const filename = book.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Book.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: 'Deleted!' });
+            })
+            .catch((error) => {
+              res.status(401).json({ error });
+            });
+        });
+      }
     })
     .catch((error) => {
       res.status(500).json({ error });
